More interaction with Zookeeper.  Workers now register themselves. Still need to implement the flatfile generation.

+++ h2o-zookeeper/src/main/java/water/zookeeper/nodes/WorkerPayload.java
@@ -3,5 +3,5 @@ package water.zookeeper.nodes;
 public class WorkerPayload extends AbstractPayload {
   public String ip;
   public int port;
-  public int pid;
+  public long pid;
 }
